Decompile rotate.s to shadow.c (#171)
* Begin to decompile rotate.s

Finished func_802CC7A0 with a possible minor error on the first
line of that function.

* First OK

Decompiled up to func_802CCF0C(). In fact, I also decompiled
func_802CD040(), but that exceeds the file boundary. I will
(obviously) include it in the next commit.

* Second OK

Up to func_802CD094().

* Third OK; add helpful struct

Decompiled up to func_802CDA44(), though I have some untested
decomp'd functions commented out for now.

* Fix git issue

* Actually fix git issue

* Fourth OK; correct many types

Through func_802CDEF0(). I am beginning to understand what these
functions are doing now. Most pointers are to "UnkRotate"
structs, vertex lists, or graphic node lists. Hopefully the rest
of the file makes it clear.

* Fifth OK

* Sixth OK

Up to func_802CE48C(). I have not yet successfully been able to
remove the goto in that function...

* Seventh OK

Up to func_802CE5A0().

* End reliance on assembly; still need to fix some constants

* Final OK; still need to clean up and document

* Add some documentation (WIP)

* Continue to add documentation

* Rename rotate to shadow throughout

* Document almost all of shadow.c

I'm not QUITE done. There's still a little documentation and
cleanup to do. For example, I'd like to un-extern all the function
declarations, tweak a few of the struct member names and local
variable names, name the remaining unnamed functions, and do
other minor style cleanup.

* Complete documentation of shadow.c

Obviously nothing is ever complete, because there are still a
couple of undocumented functions and flags, but I'm happy with
this for now.

+++ src/obj_behaviors.h
@@ -157,7 +157,7 @@ void ButterflyFollowMarioLoop(void);
 void ButterflyReturnHomeLoop(void);
 void BehButterflyLoop(void);
 void BehHootInit(void);
-f32 func_802E88EC(s32 **arg0, f32 arg1);
+f32 func_802E88EC(f32 **arg0, f32 arg1);
 // extern ? func_802E89B0(?);
 // extern ? func_802E8B3C(?);
 // extern ? func_802E8DE4(?);
